+++ app/specific/Main.js
@@ -26,8 +26,8 @@ var Main_isDebug = false;
 var Main_stringVersion = '3.0';
 var Main_stringVersion_Min = '.296';
 var Main_version_java = 296;//Always update (+1 to current value) Main_version_java after update Main_stringVersion_Min or a major update of the apk is released
-var Main_minversion = 'January 21 2020';
-var Main_version_web = 565;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
+var Main_minversion = 'January 22 2020';
+var Main_version_web = 566;//Always update (+1 to current value) Main_version_web after update Main_minversion or a major update of the web part of the app
 var Main_versionTag = Main_stringVersion + Main_stringVersion_Min + '-' + Main_minversion;
 
 var Main_cursorYAddFocus = -1;
@@ -635,6 +635,15 @@ function Main_Changelog() {
         STR_DIV_LINK + STR_ABOUT_CHANGELOG + '</div><br><br>';
 
     var changelogObj = [
+
+        {
+            title: "Web Version January 22 2020",
+            changes: [
+                "Add new setting options 'Use round channel images' (disabled by default) to settings -> Interface customization's, color style, animations and related",
+                "Improve live channels side panel looks",
+                "General improves and bug fixes"
+            ]
+        },
         {
             title: "Web Version January 21 2020",
             changes: [