+++ apk/Changelog.md
@@ -9,6 +9,13 @@ The App is divided in two part, **Apk** and **Web** you can check it version in 
 
 The app has a building update check function, so whenever there is a update a red text will show up bellow the main pages clock, also a short warning will pop on the screen let you know if is a **Apk** or **Web** update.
 
+Web Version January 22 2020
+==============
+
+* Add new setting options 'Use round channel images' (disabled by default) to settings -> Interface customization's, color style, animations and related
+* Improve live channels side panel looks
+* General improves and bug fixes
+
 Web Version January 21 2020
 ==============